chore: rename methods, move files into proper package, and remove unnecessary methods
